% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{Gapminder data}
\format{This (and the next) is a new tag that applies for data documentation.}
\source{
Note that you shouldn't use the `@export` tag when documenting data!
}
\usage{
gapminder
}
\description{
To be used to express the functionality hw09 package
}
\keyword{datasets}
