% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmfun.R
\name{lmfun}
\alias{lmfun}
\title{Display linear model objects with ease}
\usage{
lmfun(x, y, mydata)
}
\arguments{
\item{x}{The predictor variable.}

\item{y}{The dependent variable.}

\item{mydata}{The dataset to be referenced}
}
\value{
A table and qplot \code{x}.
}
\description{
This function simply prints the results of a simple linear regression
in the form of a table of regression coefficients and qplot.
}
\details{
This function is silly in that it relies heavily on functions that
already exist and live happily in other packages. However, it allows
for efficient comparison of regression coefficients, p values and
plots.

A list in this section may be excessive, but it is possible:
\itemize{
     \item This \code{pow} function is quite simple to use.
     \item No need to beat the dead horse, \code{pow}.
}
}
\examples{
lmfun(x= gapminder$lifeExp, y = gapminder$gdpPercap, mydata = gapminder)

}
