% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Raise a vector to the nth power}
\usage{
pow(x, a)
}
\arguments{
\item{x}{The vector to be raised to the nth power.}

\item{a}{The vector x is to be raised to.}
}
\value{
A vector that is the output of \code{x}.
}
\description{
That's it -- this function simply raises a vector to the nth power.
}
\details{
This function isn't complicated so there is no need to make this "details"
section lengthy


A list in this section may be excessive, but it is possible:
\itemize{
     \item This \code{pow} function is quite simple to use.
     \item No need to beat the dead horse, \code{pow}.
}
}
\examples{
pow(1:10,4)
pow(-5,4)

}
